Adding .aws

+++ dotfiles/copier.py
@@ -45,8 +45,7 @@ def _process_file(
     source_file_path = Path(os.path.join(source_folder, os.path.basename(file.filename))).absolute()
 
     exists = os.path.exists(source_file_path)
-    is_symlink = os.path.islink(source_file_path)
-    if exists and not is_symlink:
+    if exists:
         destination_file_path = Path(
             os.path.join(
                 destination_folder,
@@ -56,39 +55,29 @@ def _process_file(
 
         if os.path.exists(destination_file_path):
             if overwrite:
-                _remove_file(destination_file_path=destination_file_path, log=log, test=test)
-                _copy_file(
-                    source_file_path=source_file_path, destination_file_path=destination_file_path, log=log, test=test
-                )
+                _copy(source_file_path=source_file_path, destination_file_path=destination_file_path, log=log, test=test)
             else:
                 if log:
-                    print(f"{file} exists and overwrite is False")
+                    print(f"{file.filename} exists and overwrite is False")
         else:
-            _copy_file(source_file_path=source_file_path, destination_file_path=destination_file_path, log=log, test=test)
+            _copy(source_file_path=source_file_path, destination_file_path=destination_file_path, log=log, test=test)
     else:
         if log:
             if not exists:
                 print(f"{source_file_path} does not exist")
-            elif is_symlink:
-                print(f"{source_file_path} is a symlink")
-            else:
-                print(f"{source_file_path} had processing errors. exists: {exists}, is_symlink: {is_symlink}")
 
 
-def _remove_file(destination_file_path: str | Path, log: bool = False, test: bool = False):
-    if log:
-        print(f"Removing file: {destination_file_path}")
-    if not test:
-        os.remove(destination_file_path)
-    else:
-        print("TEST: Not removing file")
+def _copy(source_file_path: str | Path, destination_file_path: str | Path, log: bool = False, test: bool = False):
+    # checks if path is a file
+    is_file = os.path.isfile(source_file_path)
 
-
-def _copy_file(source_file_path: str | Path, destination_file_path: str | Path, log: bool = False, test: bool = False):
     if log:
         print(f"Copying {source_file_path} to {destination_file_path}")
     if not test:
-        shutil.copy(source_file_path, destination_file_path)
+        if is_file:
+            shutil.copy(source_file_path, destination_file_path)
+        else:
+            shutil.copytree(source_file_path, destination_file_path)
         if log:
             print(f"Successfully copied {source_file_path} to {destination_file_path}")
     else:
